Added slotted=runtime to apoc.cypher.runMany
